+++ tls/s2n_psk.h
@@ -88,4 +88,4 @@ int s2n_psk_free(struct s2n_psk **psk);
 int s2n_psk_set_identity(struct s2n_psk *psk, const uint8_t *identity, uint16_t identity_size);
 int s2n_psk_set_secret(struct s2n_psk *psk, const uint8_t *secret, uint16_t secret_size);
 int s2n_psk_set_hmac(struct s2n_psk *psk, s2n_psk_hmac hmac);
-int s2n_connection_set_external_psks(struct s2n_connection *conn, struct s2n_psk **psk_list, uint16_t psk_list_len);
+int s2n_connection_set_external_psks(struct s2n_connection *conn, struct s2n_psk **psk_list, uint32_t psk_list_len);
